requirements.txt
Django==4.2.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0